Barrier Works
Barrier works how I wanted it too, maybe I will add a way to deconstruct it too.

Future implementation ideas could be fixing the character for a short period of time while the barrier is "built" .

Also there is some unnecessary code and lack of comments, I need to clean it up a bit and add comments.